refactor: 페이지별 js 폴더 분리
