+++ week1/assign1/index.html
@@ -57,9 +57,7 @@
       <p>첫 번째 과제! 고생하셨씁니다 ~~</p>
     </footer>
 
-    <script type="module" src="./js/index.js"></script>
-    <script type="module" src="./js/sideNav.js"></script>
-    <script type="module" src="./js/filter.js"></script>
-    <script type="module" src="./js/banner.js"></script>
+    <script type="module" src="./js/home/index.js"></script>
+
   </body>
 </html>
