+++ week1/assign1/js/constants/const.js
