+++ week1/assign1/shoppingList.html
@@ -34,7 +34,7 @@
 
     <!-- 메인 -->
     <main class="shopping_list__container">
-        <table class="">
+        <table>
           <colgroup>
             <col>
             <col>
@@ -45,7 +45,7 @@
           <thead>
             <tr>
               <th>
-                <input type="checkbox" />
+                <input type="checkbox" id="selectAllCheckbox" />
               </th>
               <th>상품정보</th>
               <th>상품금액</th>
@@ -74,7 +74,6 @@
         </div>
     </main>
 
-    <script type="module" src="./js/sideNav.js"></script>
-    <script type="module" src="./js/getCartList.js"></script>
+    <script type="module" src="./js/shoppingList/index.js"></script>
   </body>
 </html>
